version-0.1
This is the version of the app that has the absolute minimum functionality. youre able to add and take away from the total savings. the dept and dept payment feature dont work yet, also, its possible to have a negative total savings, which is bad. And for now the Goal is hard coded to "1000000".