Connected database to UID
Now finally each user has their own table on the database as opposed to the one I used for testing, I did other things but Im so tired from coding for hours that I dont remember, not that anyones gonna read this anyway exept for my future self looking back at my first year and how foolish I am. If thats the case then Hi, and youre welcome for me putting so much work so that you could be so cool and based in the future, remember to take a break and eat some yummy food, see yha.